core/state_machine: move state_machine to its own file
