+++ core/error.rs
@@ -63,6 +63,12 @@ pub enum LimboError {
     Busy,
     #[error("Conflict: {0}")]
     Conflict(String),
+    #[error("Transaction terminated")]
+    TxTerminated,
+    #[error("Write-write conflict")]
+    WriteWriteConflict,
+    #[error("No such transaction ID: {0}")]
+    NoSuchTransactionID(String),
 }
 
 #[macro_export]
